# apps/web/src/api/specs/java-openapi.json
{
  "openapi" : "3.0.1",
  "info" : {
    "title" : "OpenAPI definition",
    "version" : "v0"
  },
  "paths" : {
    "/internal/level" : {
      "post" : {
        "tags" : [ "internal-level-controller" ],
        "operationId" : "setLevel",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/LevelRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/EnvelopeLong"
                }
              }
            }
          }
        }
      }
    },
    "/auth/register" : {
      "post" : {
        "tags" : [ "auth-controller" ],
        "operationId" : "register",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/RegisterRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/EnvelopeTokenResponse"
                }
              }
            }
          }
        }
      }
    },
    "/auth/refresh" : {
      "post" : {
        "tags" : [ "auth-controller" ],
        "operationId" : "refresh",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/RefreshRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/EnvelopeTokenResponse"
                }
              }
            }
          }
        }
      }
    },
    "/auth/login" : {
      "post" : {
        "tags" : [ "auth-controller" ],
        "operationId" : "login",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/LoginRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/EnvelopeTokenResponse"
                }
              }
            }
          }
        }
      }
    },
    "/auth/me" : {
      "get" : {
        "tags" : [ "auth-controller" ],
        "operationId" : "me",
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/EnvelopeMeView"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/ping" : {
      "get" : {
        "tags" : [ "ping-controller" ],
        "operationId" : "ping",
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/EnvelopePingData"
                }
              }
            }
          }
        }
      }
    }
  },
  "components" : {
    "schemas" : {
      "LevelRequest" : {
        "required" : [ "level", "userId" ],
        "type" : "object",
        "properties" : {
          "userId" : {
            "type" : "integer",
            "format" : "int64"
          },
          "level" : {
            "type" : "string"
          }
        }
      },
      "EnvelopeLong" : {
        "type" : "object",
        "properties" : {
          "code" : {
            "type" : "integer",
            "format" : "int32"
          },
          "message" : {
            "type" : "string"
          },
          "data" : {
            "type" : "integer",
            "format" : "int64"
          }
        }
      },
      "RegisterRequest" : {
        "required" : [ "nickname", "password", "username" ],
        "type" : "object",
        "properties" : {
          "username" : {
            "maxLength" : 64,
            "minLength" : 0,
            "type" : "string"
          },
          "email" : {
            "maxLength" : 254,
            "minLength" : 0,
            "type" : "string"
          },
          "password" : {
            "maxLength" : 72,
            "minLength" : 8,
            "type" : "string"
          },
          "nickname" : {
            "maxLength" : 64,
            "minLength" : 0,
            "type" : "string"
          },
          "ageGroup" : {
            "maxLength" : 16,
            "minLength" : 0,
            "type" : "string"
          }
        }
      },
      "EnvelopeTokenResponse" : {
        "type" : "object",
        "properties" : {
          "code" : {
            "type" : "integer",
            "format" : "int32"
          },
          "message" : {
            "type" : "string"
          },
          "data" : {
            "$ref" : "#/components/schemas/TokenResponse"
          }
        }
      },
      "TokenResponse" : {
        "type" : "object",
        "properties" : {
          "accessToken" : {
            "type" : "string"
          },
          "refreshToken" : {
            "type" : "string"
          },
          "expiresIn" : {
            "type" : "integer",
            "format" : "int64"
          },
          "userId" : {
            "type" : "integer",
            "format" : "int64"
          }
        }
      },
      "RefreshRequest" : {
        "required" : [ "refreshToken" ],
        "type" : "object",
        "properties" : {
          "refreshToken" : {
            "type" : "string"
          }
        }
      },
      "LoginRequest" : {
        "required" : [ "password", "username" ],
        "type" : "object",
        "properties" : {
          "username" : {
            "type" : "string"
          },
          "password" : {
            "type" : "string"
          }
        }
      },
      "EnvelopeMeView" : {
        "type" : "object",
        "properties" : {
          "code" : {
            "type" : "integer",
            "format" : "int32"
          },
          "message" : {
            "type" : "string"
          },
          "data" : {
            "$ref" : "#/components/schemas/MeView"
          }
        }
      },
      "MeView" : {
        "type" : "object",
        "properties" : {
          "userId" : {
            "type" : "integer",
            "format" : "int64"
          },
          "username" : {
            "type" : "string"
          },
          "nickname" : {
            "type" : "string"
          },
          "level" : {
            "type" : "string"
          }
        }
      },
      "EnvelopePingData" : {
        "type" : "object",
        "properties" : {
          "code" : {
            "type" : "integer",
            "format" : "int32"
          },
          "message" : {
            "type" : "string"
          },
          "data" : {
            "$ref" : "#/components/schemas/PingData"
          }
        }
      },
      "PingData" : {
        "type" : "object",
        "properties" : {
          "status" : {
            "type" : "string"
          },
          "service" : {
            "type" : "string"
          }
        }
      }
    }
  }
}